Do not include machine id into the firewall find query.
It is sufficient to know that an arbitrary firewall exists for the cluster, then the algorithm does not need to create a new one. Even if the Gardener is not aware of the correct machine id of the firewall. Otherwise, there is the potential that multiple firewalls for the same cluster will be created, which leads to very confusing network behavior.